rename project dir, ignore build directory
