+++ .gitignore
@@ -3,3 +3,6 @@
 
 # Cache
 __pycache__/
+
+# Build
+dist/
